# docs/instant-messaging-api-spec.yml
openapi: 3.0.3
info:
  title: Instant Messaging API
  description: API for managing messaging channels
  version: 0.1.0
  # Antonio Paulino - 50512
  # Bernardo Pereira - 50493
servers:
  - url: http://localhost:8080/api
    description: Local server
...
paths:
...
  /api/channels:
    post:
      tags:
        - Channels
      summary: Create a new channel
      security:
        - bearerAuth: [ ]
      requestBody:
        description: Details to create a new channel, either public or private
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelCreationInputModel'
      responses:
        '201':
          description: Channel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelCreationOutputModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    get:
        tags:
            - Channels
        summary: Get a list of all channels
        security:
          - bearerAuth: []
        parameters:
          - $ref: '#/components/parameters/Offset'
          - $ref: '#/components/parameters/Limit'
          - $ref: '#/components/parameters/GetCount'
          - $ref: '#/components/parameters/Sort'
          - $ref: '#/components/parameters/Name'
          - name: sortBy
            in: query
            required: false
            schema:
                type: string
                enum: [ id, name, createdAt ]
                default: name
        responses:
          '200':
            description: List of channels
            content:
              application/json:
                schema:
                    $ref: '#/components/schemas/ChannelsPaginatedOutputModel'
          '401':
              $ref: '#/components/responses/Unauthorized'
          '429':
            $ref: '#/components/responses/TooManyRequests'
  /api/channels/{channelID}:
    get:
      tags:
        - Channels
      summary: Get channel details by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
      responses:
        '200':
          description: Details of the channel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelOutputModel'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    put:
      tags:
        - Channels
      summary: Update the channel's details
      security:
          - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
      requestBody:
        description: New channel details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelUpdateInputModel'
      responses:
        '204':
          description: Channel updated successfully
        '400':
            $ref: '#/components/responses/BadRequest'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    delete:
      tags:
        - Channels
      summary: Delete a channel
      security:
          - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
      responses:
        '204':
          description: Channel deleted successfully
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/channels/{channelID}/members/{userID}:
    put:
      tags:
        - Channels
      summary: Join a public channel
      security:
          - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
        - $ref: '#/components/parameters/UserID'
      responses:
        '201':
          description: User added successfully
        '400':
            $ref: '#/components/responses/BadRequest'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    delete:
      tags:
        - Channels
      summary: Remove a user from a channel
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
        - $ref: '#/components/parameters/UserID'
      responses:
        '204':
          description: User left successfully
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    patch:
        tags:
            - Channels
        summary: Update a user's role in a channel
        security:
            - bearerAuth: []
        parameters:
            - $ref: '#/components/parameters/ChannelID'
            - $ref: '#/components/parameters/UserID'
        requestBody:
            description: New role
            required: true
            content:
              application/json:
                  schema:
                    $ref: '#/components/schemas/RoleInputModel'
        responses:
            '204':
                description: User role updated successfully
            '400':
                $ref: '#/components/responses/BadRequest'
            '401':
                $ref: '#/components/responses/Unauthorized'
            '403':
                $ref: '#/components/responses/Forbidden'
            '404':
                $ref: '#/components/responses/NotFound'
            '429':
              $ref: '#/components/responses/TooManyRequests'
...
components:
  schemas:
...
    UserOutputModel:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
...
    PaginationOutputModel:
      type: object
      properties:
        current:
          type: integer
        next:
          type: integer
        previous:
          type: integer
        totalPages:
          type: integer
        total:
          type: integer
    ChannelCreationInputModel:
      type: object
      properties:
        name:
          type: string
        defaultRole:
          type: string
          enum: [ member, guest ]
        isPublic:
          type: boolean
      required:
        - name
        - defaultRole
        - isPublic
    ChannelOutputModel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        isPublic:
          type: boolean
        owner:
          $ref: '#/components/schemas/UserOutputModel'
        members:
          type: array
          items:
            $ref: '#/components/schemas/ChannelMemberOutputModel'
        createdAt:
          type: string
          format: date-time
    ChannelMemberOutputModel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
          enum: [ owner, member, guest ]
    ChannelsPaginatedOutputModel:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: '#/components/schemas/ChannelOutputModel'
        pagination:
          $ref: '#/components/schemas/PaginationOutputModel'
...
    Problem:
      description: Error response
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          format: uri
        status:
          type: integer
        detail:
          type: string
    InputValidationProblem:
      description: Validation error for Body, Query, or Path parameters
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          format: uri
        status:
          type: integer
        detail:
          type: string
        errors:
          type: array
          items:
            type: string
    RoleInputModel:
      type: object
      properties:
        role:
          type: string
          enum: [ owner, member, guest ]
      required:
        - role
    ChannelUpdateInputModel:
      type: object
      properties:
        name:
          type: string
        isPublic:
          type: boolean
      required:
        - name
        - isPublic
    ChannelCreationOutputModel:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Problem'
              - $ref: '#/components/schemas/InputValidationProblem'
    Unauthorized:
        description: Unauthorized
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Problem'
    Forbidden:
        description: Forbidden
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Problem'
    TooManyRequests:
        description: Too many requests
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Problem'
    NotFound:
        description: Not found
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Problem'
...
  parameters:
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        maximum: 100
        minimum: 1
    Name:
      name: name
      in: query
      required: false
      schema:
        type: string
    GetCount:
      name: getCount
      in: query
      required: false
      schema:
        type: boolean
        default: true
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: [ asc, desc ]
        default: asc
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: number
    ChannelID:
      name: channelID
      in: path
      required: true
      schema:
        type: number
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID